mini_arabic_co_pilot/requirements.txt
# Core FastAPI and web dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Streamlit frontend
streamlit==1.28.1

# OCR and Image Processing
easyocr==1.7.0
Pillow==10.1.0
opencv-python==4.8.1.78

# Machine Learning and NLP
torch==2.1.0
transformers==4.35.2
sentence-transformers==2.2.2
scikit-learn==1.3.2

# PDF Processing
PyMuPDF==1.23.8
pdfplumber==0.10.3

# Vector Database and Search
faiss-cpu==1.7.4

# Time Series Forecasting
prophet==1.1.4
statsmodels==0.14.0

# Data Processing
pandas==2.1.3
numpy==1.25.2

# Visualization
plotly==5.17.0
matplotlib==3.8.2

# HTTP requests
requests==2.31.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Optional: MLflow for experiment tracking
# mlflow==2.8.1

# Optional: Docker support
# docker==6.1.3